fixShop
